refactor(fdb): extract VLAN-community sub-engine to vlan_community.go (#151)
